+++ Lyric Finder/Lyric Finder/Models/IIISong.m
@@ -13,7 +13,7 @@
 - (instancetype)initWithTitle:(NSString *)title
                        artist:(NSString *)artist
                        lyrics:(NSString *)lyrics
-                       rating:(NSInteger)rating
+                       rating:(int)rating
 {
     self = [super init];
     if (self) {
